quelques modif

+++ src/main/java/com/inventaire/Inventaire_Actifs/controllers/User_Controller.java
@@ -16,57 +16,6 @@ import java.util.stream.Collectors;
 @CrossOrigin(origins = "http://localhost:4200")
 @RestController
 @RequestMapping("/api/users")
-//public class User_Controller {
-//
-//    @Autowired
-//    private User_Service userService;
-//
-//    @GetMapping
-//    public ResponseEntity<List<UserDTO>> getAllUsers() {
-//        List<User> users = userService.getAllUsers();
-//        List<UserDTO> usersDTO = users.stream().map(UserDTO::new).collect(Collectors.toList());
-//        return new ResponseEntity<>(usersDTO, HttpStatus.OK);
-//    }
-//
-//    @GetMapping("/{id}")
-//    public ResponseEntity<User> getUserById(@PathVariable Long id) {
-//        User user = userService.getUserById(id);
-//        return new ResponseEntity<>(user, HttpStatus.OK);
-//    }
-//
-//    @PostMapping("/signup")
-//    public ResponseEntity<User> signUp(@RequestBody User user) {
-//        User createdUser = userService.createUser(user);
-//        return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
-//    }
-//
-//
-//
-//
-//    @PutMapping("/{id}")
-//    public ResponseEntity<Void> updateUser(@PathVariable Long id, @RequestBody User user) {
-//        User updatedUser = userService.updateUser(id, user);
-//        if (updatedUser != null) {
-//            return ResponseEntity.noContent().build();
-//        } else {
-//            return ResponseEntity.notFound().build();
-//        }
-//    }
-//
-//
-//    @DeleteMapping("/{id}")
-//    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-//        boolean deleted = userService.deleteUser(id);
-//        if (deleted) {
-//            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-//        } else {
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//        }
-//    }
-//
-//
-//
-//}
 
 
 public class User_Controller {
@@ -85,7 +34,7 @@ public class User_Controller {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id) {
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Integer id) {
         UserDTO user = userService.getUserById(id);
         if (user != null) {
             return new ResponseEntity<>(user, HttpStatus.OK);
@@ -101,7 +50,7 @@ public class User_Controller {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<UserDTO> updateUser(@PathVariable Long id, @RequestBody UserDTO userDTO) {
+    public ResponseEntity<UserDTO> updateUser(@PathVariable Integer id, @RequestBody UserDTO userDTO) {
         UserDTO updatedUser = userService.updateUser(id, userDTO);
         if (updatedUser != null) {
             return new ResponseEntity<>(updatedUser, HttpStatus.OK);
@@ -111,7 +60,7 @@ public class User_Controller {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
         boolean deleted = userService.deleteUser(id);
         if (deleted) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
